pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.77.0",
    "plotly>=6.0.1",
    "streamlit>=1.45.0",
    "python-dotenv>=1.0.0",
]
